+++ bin/load.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+#
+# This script runs within the CircleCI environment to build stackstorm images.
+
+set -euo pipefail
+IDS=$'\n\t'
+
+source bin/common.sh
+
+for name in stackstorm; do
+  # Load the tarball (tags are automatically loaded)
+  ${dry_run} docker load -i /workspace/tar/${name}.tar
+done