+++ CHANGELOG.rst
@@ -4,6 +4,14 @@ Changelog
 Current
 ---
 
+2018-02-27
+----------
+
+Changed
+~~~~~~~
+
+* Pin DB's to specific, tested versions.
+
 2017-10-23
 ----------